refactor: optimize
