+++ app/components/UI/Bridge/hooks/useTokens.ts
@@ -8,7 +8,7 @@ import {
 } from '@metamask/bridge-controller';
 import { zeroAddress } from 'ethereumjs-util';
 import { POLYGON_NATIVE_TOKEN } from '../constants/assets';
-import { includeOnlyTradableTokens } from '../utils/includeOnlyTradableTokens';
+import { isTradableToken } from '../utils/isTradableToken';
 
 interface UseTokensProps {
   topTokensChainId?: Hex | CaipChainId;
@@ -78,19 +78,25 @@ export function useTokens({
   const tokensWithoutBalance = (topTokens ?? [])
     .concat(remainingTokens ?? [])
     .filter((token) => {
+      if (!isTradableToken(token)) {
+        return false;
+      }
+
       const tokenKey = getTokenKey(token);
       return !tokensWithBalanceSet.has(tokenKey);
-    })
-    .filter(includeOnlyTradableTokens);
+    });
 
   // Combine tokens with balance and filtered tokens and filter out excluded tokens
   const allTokens = tokensWithBalance
     .concat(tokensWithoutBalance)
     .filter((token) => {
+      if (!isTradableToken(token)) {
+        return false;
+      }
+
       const tokenKey = getTokenKey(token);
       return !excludedTokensSet.has(tokenKey);
-    })
-    .filter(includeOnlyTradableTokens);
+    });
 
   const tokensToRender = tokensWithBalance
     .concat(
@@ -100,10 +106,13 @@ export function useTokens({
       }) ?? [],
     )
     .filter((token) => {
+      if (!isTradableToken(token)) {
+        return false;
+      }
+
       const tokenKey = getTokenKey(token);
       return !excludedTokensSet.has(tokenKey);
-    })
-    .filter(includeOnlyTradableTokens);
+    });
 
   return { allTokens, tokensToRender, pending };
 }
